backend/requirements.txt
dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-allauth
pillow 
psycopg2
requests